Return timezone aware objects whenever zulu time is specified
This relates to the fix for issue #175, commit 7dfbe07, which
accidentally changes the datetime objects from python-metar
from timezone aware to non-timezone-aware.

This change restores timezone awareness in datetime objects as long
as the time in the METAR is specified with a trailing Z. This makes
it easier to make robust comparisons of datetimes when using this
library.